refactor: move console.log silencing into versioner
