+++ package.json
@@ -23,7 +23,7 @@
     "test:sigs": "./test/signatures/test-sigs.bash",
     "test:storage": "./test/storage/test-storage.bash",
     "test:docker": "./test/docker/test-docker.bash",
-    "orbit:contracts:version": "hardhat run scripts/orbit-versioner/orbitVersionerCaller.ts",
+    "orbit:contracts:version": "hardhat run scripts/orbit-versioner/orbitVersioner.ts",
     "gas-snapshot": "forge snapshot --match-path \"test/unit/**/*.t.sol\"",
     "fix": "yarn format; yarn test:sigs; yarn test:storage; yarn gas-snapshot"
   },
